refactor: 상품 아이템 컴포넌트, 컨테이너 폴더 구조 변경
