+++ src/component/ProductList/ProductContainer/index.jsx
@@ -1,4 +1,4 @@
-import Product from 'component/ProductList/Product';
+import Product from 'component/ProductList/ProductContainer/Product';
 import { PATH } from 'constant/path';
 import { useDispatch } from 'react-redux';
 import { useNavigate } from 'react-router-dom';
@@ -6,12 +6,11 @@ import { selectProduct } from 'store/action/selectActions';
 import { addCartProductThunk } from 'store/thunk/productThunk';
 
 function ProductContainer({ product }) {
-  const { image, name, price, id } = product;
   const dispatch = useDispatch();
   const navigate = useNavigate();
 
-  const handleProductClick = productId => {
-    dispatch(selectProduct(productId));
+  const handleProductClick = () => {
+    dispatch(selectProduct(product.id));
     navigate(PATH.PRODUCT_DETAIL);
   };
 
@@ -23,10 +22,7 @@ function ProductContainer({ product }) {
     <Product
       handleProductClick={handleProductClick}
       handleCartClick={handleCartClick}
-      image={image}
-      name={name}
-      price={price}
-      id={id}
+      product={product}
     ></Product>
   );
 }
